use real time as min clock

+++ fantoch_ps/src/protocol/atlas.rs
@@ -42,7 +42,7 @@ impl<KC: KeyClocks> Protocol for Atlas<KC> {
     fn new(
         process_id: ProcessId,
         config: Config,
-    ) -> (Self, Vec<(PeriodicEvent, usize)>) {
+    ) -> (Self, Vec<(PeriodicEvent, u64)>) {
         // compute fast and write quorum sizes
         let (fast_quorum_size, write_quorum_size) = config.atlas_quorum_sizes();
 
@@ -71,7 +71,7 @@ impl<KC: KeyClocks> Protocol for Atlas<KC> {
         };
 
         // create periodic events
-        let gc_delay = config.garbage_collection_interval();
+        let gc_delay = config.garbage_collection_interval() as u64;
         let events = vec![(PeriodicEvent::GarbageCollection, gc_delay)];
 
         // return both
