+++ fantoch_ps/src/protocol/fpaxos.rs
@@ -35,7 +35,7 @@ impl Protocol for FPaxos {
     fn new(
         process_id: ProcessId,
         config: Config,
-    ) -> (Self, Vec<(Self::PeriodicEvent, usize)>) {
+    ) -> (Self, Vec<(Self::PeriodicEvent, u64)>) {
         // compute fast and write quorum sizes
         let fast_quorum_size = 0; // there's no fast quorum as we don't have fast paths
         let write_quorum_size = config.fpaxos_quorum_size();
@@ -67,7 +67,7 @@ impl Protocol for FPaxos {
         };
 
         // create periodic events
-        let gc_delay = config.garbage_collection_interval();
+        let gc_delay = config.garbage_collection_interval() as u64;
         let events = vec![(PeriodicEvent::GarbageCollection, gc_delay)];
 
         // return both
